+++ src/utils/api.tsx
@@ -74,6 +74,7 @@ export default class API {
 			.filter( (c: Course) => {
 				return c.semester === this.SEMESTER
 			})
+		console.log(courses)
 		return courses.sort(Course.compare)
 	}